Make sum returns a nullable value
From PostgreSQL documentation:

> It should be noted that except for count, these functions return a null value
when no rows are selected. In particular, sum of no rows returns null, not zero
as one might expect, and array_agg returns null rather than an empty array when
there are no input rows. The coalesce function can be used to substitute zero or
an empty array for null when necessary.

So the `sum` aggregate should not return a `NotNull` column.