parameter check
I wish there was a way to do this at the type level, but I cannot think of any. Right now, Squeal doesn't stop you from writing queries that don't use all of their parameters, even though this is a SQL error. Moreover, the SQL error itself is not so helpful.
```
*** Exception: PQException {sqlExecStatus = FatalError, sqlStateCode = Just "42P18", sqlErrorMessage = Just "ERROR:  could not determine data type of parameter $2\n"}
```
I'm not sure that this is any better though. But putting it up for some comments.