making cross schema foreign keys backwards compatible
In Squeal 0.6 the `ForeignKey` constuctor has three arguments

```Haskell
ForeignKey [Symbol] Symbol [Symbol]
```

The second argument was the table being referenced; but there was a bug in cross-schema support for doing a `createTable` with a cross-schema `foreignKey`.

The obvious solution to add a fourth argument with the schema being referenced. But that causes a backward compatibility problem since users will have to update their schema definitions. Also, it causes a problem for squealgen because it then has to be squeal version aware.

So, instead we'll do what squealgen is _already_ doing. for intra-schema references, just use the table name, and for cross-schema references, use the schema qualified table name. Now it's automatically backwards compatible (except for enabling cross-schema foreign keys) and automatically compatible with squealgen. To do that, we define a type family `References` that computes the string for either intra- or cross-schema references.

```Haskell
type family References (sch :: Symbol) (tab :: Symbol) (sch0 :: Symbol) where
  References sch tab sch = tab
  References sch tab _   = sch `AppendSymbol` "." `AppendSymbol` tab
```